+++ packages/picasso/src/Input/styles.ts
@@ -21,26 +21,5 @@ export default ({ palette }: Theme) =>
     },
     inputMultilineResizable: {
       resize: 'vertical'
-    },
-    icon: {
-      flex: '1 1 0%' // fix for IE11
-    },
-    limiterLabel: {
-      color: palette.grey.main2,
-      fontSize: '0.625rem',
-      lineHeight: '1rem'
-    },
-    limiterLabelError: {
-      color: palette.red.main
-    },
-    limiterMultiline: {
-      position: 'absolute',
-      bottom: 0,
-      width: 'calc(100% - 1.25rem)',
-      height: '1.25rem',
-      justifyContent: 'flex-start',
-      margin: 0,
-      padding: '0.25rem 0',
-      borderTop: `1px solid ${palette.grey.lighter2}`
     }
   })
