+++ packages/picasso/src/InputIconAdornment/index.ts
@@ -1,0 +1,6 @@
+import { OmitInternalProps } from '@toptal/picasso-shared'
+
+import { Props } from './InputIconAdornment'
+
+export { default } from './InputIconAdornment'
+export type InputIconAdornmentProps = OmitInternalProps<Props>
