+++ packages/picasso/src/InputLimitAdornment/index.ts
@@ -1,0 +1,6 @@
+import { OmitInternalProps } from '@toptal/picasso-shared'
+
+import { Props } from './InputLimitAdornment'
+
+export { default } from './InputLimitAdornment'
+export type InputLimitAdornmentProps = OmitInternalProps<Props>
